refactor(mergegate): mergegate.py をモジュール分割（954行→329行）
Phase A: guard 関数群を mergegate_guards.py に抽出
- MergeGateError, _check_worktree_guard, _check_worker_window_guard
- _check_running_guard, _check_phase_review_guard
- _state_read, _state_write, _board_update, _detect_repo_mode

Phase B: 内部メソッドを MergeGateOperationsMixin として mergegate_ops.py に抽出
- _get_issue_labels, _gh_issue_state, _verify_and_close_issue
- _ensure_closes_link, _check_base_drift
- _check_deps_yaml_conflict_and_rebase, _run_merge, _post_merge_cleanup
- _remove_worktree, _delete_remote_branch, _kill_worker_window
- _now_iso, _detect_autopilot_dir, _detect_scripts_root

deps.yaml に autopilot-mergegate-guards / autopilot-mergegate-ops を追加。
テスト: _check_phase_review_guard の mock 漏れ（pre-existing bug）を修正。
77 tests pass。

Closes #462
